backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.1.2
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.109.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
netifaces==0.11.0
numpy==1.26.4
pandas==2.1.4
passlib==1.7.4
psutil==7.2.2
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==12.0
